+++ CHANGELOG.md
@@ -2,6 +2,20 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+## [4.0.0](https://github.com/peerigon/clockodo/compare/v3.2.0...v4.0.0) (2020-01-24)
+
+
+### Bug Fixes
+
+* Change id type from string to number ([0513722](https://github.com/peerigon/clockodo/commit/0513722))
+
+
+### BREAKING CHANGES
+
+* Clockodo REST API uses type number for ids.
+
+
+
 ## [3.2.0](https://github.com/peerigon/clockodo/compare/v3.1.6...v3.2.0) (2019-12-17)